requirements.txt
flask==2.3.2
openai==0.27.2
pinecone-client==2.2.1
pandas==2.0.1
gunicorn==20.1.0
tiktoken==0.3.3
Flask==2.3.2
Flask-HTTPAuth==4.8.0
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-SQLAlchemy==3.0.3
APScheduler==3.10.4